Refatoração e Loans
- Feito uma refatoração para que fique mais organizados, agora cada
combo está em seu local com um alias/widget cadastrado.
- Loans foi alterado para que não seja mais permitido efetuar
alterações e exclusões.
- Loans agora tema a ação Renew.
- Aplicado programação defensiva para o caso de chamar uma ação de
Destroy/Renew quando não tem nenhum item selecionado (não faz nada
nesse caso, não dando erro de JS no log do console).

+++ public/app/controller/Books.js
@@ -55,16 +55,18 @@ Ext.define('AW.controller.Books', {
 			record = grid.getSelectionModel().getSelection()[0],
 			store = grid.getStore();
 
-		record.destroy({
-			success: function(response) {
-				var store = grid.getStore();
-				store.reload();
-			},
-			failure: function(response) {
-				// TODO: isolar comportamento repetitivo.
-				Ext.Msg.alert('Error', 'Ocorreu algum problema no servidor.');
-			}
-		});
+		if (record) {
+			record.destroy({
+				success: function(response) {
+					var store = grid.getStore();
+					store.reload();
+				},
+				failure: function(response) {
+					// TODO: isolar comportamento repetitivo.
+					Ext.Msg.alert('Error', 'Ocorreu algum problema no servidor.');
+				}
+			});
+		}
 	},
 
 	onEditClick: function(grid, record, item, index, e, eOpts) {
